Group previous misconduct files into a folder
