+++ app/views/referrals/previous_misconduct/check_answers/edit.html.erb
@@ -11,7 +11,7 @@
 
     <%= render PreviousMisconductComponent.new(referral: current_referral) %>
 
-    <%= form_with model: @previous_misconduct_form, url: referral_previous_misconduct_path(current_referral), method: :patch do |f| %>
+    <%= form_with model: @previous_misconduct_check_answers_form, url: [current_referral.routing_scope, current_referral, :previous_misconduct, :check_answers], method: :patch do |f| %>
       <%= f.govuk_radio_buttons_fieldset :previous_misconduct_complete, hint: { text: 'You can still make changes to a completed section' }, legend: { size: 'm', text: 'Have you completed this section?' } do %>
         <%= f.hidden_field :previous_misconduct_complete %>
         <%= f.govuk_radio_button :previous_misconduct_complete, true, label: { text: "Yes, I’ve completed this section" }, link_errors: true %>
