+++ app/views/referrals/previous_misconduct/reported/edit.html.erb
@@ -1,9 +1,9 @@
-<% content_for :page_title, "#{'Error: ' if @previous_misconduct_reported_form.errors.any?}Has there been any previous misconduct, disciplinary action or complaints?" %>
+<% content_for :page_title, "#{'Error: ' if @reported_form.errors.any?}Has there been any previous misconduct, disciplinary action or complaints?" %>
 <% content_for :back_link_url, return_to_session_or(polymorphic_path([:edit, current_referral.routing_scope, current_referral])) %>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= form_with model: @previous_misconduct_reported_form, url: [current_referral.routing_scope, current_referral, :previous_misconduct_reported], method: :patch do |f| %>
+    <%= form_with model: @reported_form, url: [current_referral.routing_scope, current_referral, :previous_misconduct_reported], method: :patch do |f| %>
       <h1 class="govuk-heading-l">Previous allegations</h1>
       <p class="govuk-body">This includes:</p>
       <ul class="govuk-list govuk-list--bullet govuk-!-margin-bottom-6">
