+++ spec/forms/referrals/previous_misconduct/check_answers_form_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe PreviousMisconductForm, type: :model do
+RSpec.describe Referrals::PreviousMisconduct::CheckAnswersForm, type: :model do
   let(:referral) { create(:referral) }
   let(:form) { described_class.new(referral:, previous_misconduct_complete:) }
   let(:previous_misconduct_complete) { true }
